requirements.txt
# OddCity - Python Dependencies
# Install: pip install -r requirements.txt

# Web Framework
Flask>=3.0.0
Flask-Session>=0.5.0
Flask-CORS>=4.0.0
Flask-Limiter>=3.5.0

# Database
mysql-connector-python>=8.2.0

# Security
Werkzeug>=3.0.0

# API Documentation
flasgger>=0.9.7

# Rate Limiting Storage (optional - for production)
# redis>=5.0.0  # Uncomment if using Redis for rate limiting




